型の修正

+++ FluentRviz.hpp
@@ -466,7 +466,7 @@ namespace internal {
         }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct PositionEnabler : PositionHelper<Marker> { };
 
     template<typename T>
@@ -483,7 +483,7 @@ namespace internal {
         }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct OrientationEnabler : OrientationHelper<Marker> { };
 
     template<typename T>
@@ -527,35 +527,35 @@ namespace internal {
         { return std::move(*this).ScaleHelper<T>::scale(0, 0, height); }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct ScaleEnabler { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ScaleEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_common_scale_available_v<MarkerType>>>
         : ScaleHelper<Marker> { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ScaleEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_arrow_marker_v<MarkerType> && is_contained_v<option::Arrow::POSE, Options...>>>
         : PoseArrowScaleHelper<Marker> { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ScaleEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_arrow_marker_v<MarkerType> && is_contained_v<option::Arrow::VECTOR, Options...>>>
         : VectorArrowScaleHelper<Marker> { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ScaleEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_points_marker_v<MarkerType>>>
         : PointScaleHelper<Marker> { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ScaleEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_line_marker_v<MarkerType>>>
         : LineScaleHelper<Marker> { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ScaleEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_text_view_facing_marker_v<MarkerType>>>
         : TextScaleHelper<Marker> { };
@@ -574,10 +574,10 @@ namespace internal {
         }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct ColorEnabler { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ColorEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<!is_colors_available_v<MarkerType> || !is_contained_v<option::Color::SEPARATE, Options...>>>
         : ColorHelper<Marker> { };
@@ -654,15 +654,15 @@ namespace internal {
         }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct PointsEnabler { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct PointsEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_points_available_v<MarkerType>>>
         : PointsHelper<Marker> { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct PointsEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_arrow_marker_v<MarkerType> && is_contained_v<option::Arrow::VECTOR, Options...>>>
         : ArrowPointsHelper<Marker> { };
@@ -685,10 +685,10 @@ namespace internal {
         }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct ColorsEnabler { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct ColorsEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_colors_available_v<MarkerType> && is_contained_v<option::Color::SEPARATE, Options...>>>
         : ColorsHelper<Marker> { };
@@ -704,10 +704,10 @@ namespace internal {
         }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct TextEnabler { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct TextEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_text_view_facing_marker_v<MarkerType>>>
         : TextHelper<Marker> { };
@@ -731,10 +731,10 @@ namespace internal {
         }
     };
 
-    template<typename Marker, size_t MarkerType, typename Option, typename Enable = void>
+    template<typename Marker, int32_t MarkerType, typename Option, typename Enable = void>
     struct MeshResourceEnabler { };
 
-    template<typename Marker, size_t MarkerType, auto... Options>
+    template<typename Marker, int32_t MarkerType, auto... Options>
     struct MeshResourceEnabler<Marker, MarkerType, OptionPack<Options...>,
         std::enable_if_t<is_mesh_resource_marker_v<MarkerType>>>
         : MeshResourceHelper<Marker> { };
